+++ src/main/java/roomescape/domain/greeting/controller/GreetingController.java
@@ -1,10 +1,10 @@
-package roomescape.controller;
+package roomescape.domain.greeting.controller;
 
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.Greeting;
+import roomescape.domain.greeting.model.Greeting;
 
 @RestController
 public class GreetingController {
