+++ src/main/java/roomescape/domain/home/controller/HomeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.domain.home.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
